Read with cv2

+++ train/build_utils/data_utils.py
@@ -22,7 +22,7 @@ class MayoDataset(torch.utils.data.dataset.Dataset):
         if self.transform is not None:
             image = self.transform(image)
 
-        label = torch.Tensor([S2I_LBL_MAP[label]])
+        label = torch.Tensor([S2I_LBL_MAP[label]]).long()
         if self.target_transform is not None:
             label = self.target_transform(label)
